+++ contracts/interfaces/external/synthetix/ISnxDepot.sol
@@ -5,33 +5,36 @@ interface ISnxDepot {
     // Views
     function fundsWallet() external view returns (address payable);
 
-    function maxEthPurchase() external view returns (uint);
+    function maxEthPurchase() external view returns (uint256);
 
-    function minimumDepositAmount() external view returns (uint);
+    function minimumDepositAmount() external view returns (uint256);
 
-    function synthsReceivedForEther(uint amount) external view returns (uint);
+    function synthsReceivedForEther(uint256 amount) external view returns (uint256);
 
-    function totalSellableDeposits() external view returns (uint);
+    function totalSellableDeposits() external view returns (uint256);
 
     // Mutative functions
-    function depositSynths(uint amount) external;
+    function depositSynths(uint256 amount) external;
 
-    function exchangeEtherForSynths() external payable returns (uint);
+    function exchangeEtherForSynths() external payable returns (uint256);
 
-    function exchangeEtherForSynthsAtRate(uint guaranteedRate) external payable returns (uint);
+    function exchangeEtherForSynthsAtRate(uint256 guaranteedRate) external payable returns (uint256);
 
     function withdrawMyDepositedSynths() external;
 
     // Note: On mainnet no SNX has been deposited. The following functions are kept alive for testnet SNX faucets.
-    function exchangeEtherForSNX() external payable returns (uint);
+    function exchangeEtherForSNX() external payable returns (uint256);
 
-    function exchangeEtherForSNXAtRate(uint guaranteedRate, uint guaranteedSynthetixRate) external payable returns (uint);
+    function exchangeEtherForSNXAtRate(uint256 guaranteedRate, uint256 guaranteedSynthetixRate)
+        external
+        payable
+        returns (uint256);
 
-    function exchangeSynthsForSNX(uint synthAmount) external returns (uint);
+    function exchangeSynthsForSNX(uint256 synthAmount) external returns (uint256);
 
-    function synthetixReceivedForEther(uint amount) external view returns (uint);
+    function synthetixReceivedForEther(uint256 amount) external view returns (uint256);
 
-    function synthetixReceivedForSynths(uint amount) external view returns (uint);
+    function synthetixReceivedForSynths(uint256 amount) external view returns (uint256);
 
-    function withdrawSynthetix(uint amount) external;
+    function withdrawSynthetix(uint256 amount) external;
 }
